LayoutTests/inspector/dom-debugger/attribute-modified-style-expected.txt

Tests functionality of attribute modified DOM breakpoints when modifying the inline style attribute.

Adding attribute modified breakpoint to #test...
PASS: Added attribute modified breakpoint to #test.

== Running test suite: DOMBreakpoints.AttributeModified.Style
-- Running test case: DOMBreakpoints.AttributeModified.Style.Add
Getting 'display'...
PASS: 'display' should have the value ''.
Setting 'display: none;'...
PASS: Paused.
Getting 'display'...
PASS: 'display' should have the value ''.
PASS: Resumed.
Getting 'display'...
PASS: 'display' should have the value 'none'.

-- Running test case: DOMBreakpoints.AttributeModified.Style.ReplaceSame
Getting 'display'...
PASS: 'display' should have the value 'none'.
Setting 'display: none;'...
PASS: Paused.
Getting 'display'...
PASS: 'display' should have the value 'none'.
PASS: Resumed.
Getting 'display'...
PASS: 'display' should have the value 'none'.

-- Running test case: DOMBreakpoints.AttributeModified.Style.ReplaceDifferent
Getting 'display'...
PASS: 'display' should have the value 'none'.
Setting 'display: block;'...
PASS: Paused.
Getting 'display'...
PASS: 'display' should have the value 'none'.
PASS: Resumed.
Getting 'display'...
PASS: 'display' should have the value 'block'.

-- Running test case: DOMBreakpoints.AttributeModified.Style.Remove
Getting 'display'...
PASS: 'display' should have the value 'block'.
Removing 'display'...
PASS: Paused.
Getting 'display'...
PASS: 'display' should have the value 'block'.
PASS: Resumed.
Getting 'display'...
PASS: 'display' should have the value ''.

-- Running test case: DOMBreakpoints.AttributeModified.Style.Text
Getting 'display'...
PASS: 'display' should have the value ''.
Setting style text to ''...
PASS: Paused.
Getting 'display'...
PASS: 'display' should have the value ''.
PASS: Resumed.
Getting 'display'...
PASS: 'display' should have the value 'none'.
